refactor: continue cleanup items that came out of the review
