+++ api/src/main/java/com/microsoft/gctoolkit/io/RotatingGCLogFile.java
@@ -49,7 +49,11 @@ public class RotatingGCLogFile extends GCLogFile {
     @Override
     public Stream<String> stream() throws IOException {
         if ( getMetaData().isDirectory() || getMetaData().isPlainText() || getMetaData().isZip())
-            return getMetaData().logFiles().flatMap(segment -> segment.stream());
+            return getMetaData().logFiles()
+                    .flatMap(segment -> segment.stream())
+                    .filter(Objects::nonNull)
+                    .map(String::trim)
+                    .filter(s -> s.length() > 0);
         else // yes, this is returning an empty stream.
             return new ArrayList<String>().stream();
     }
