+++ vertx/src/main/java/com/microsoft/gctoolkit/vertx/jvm/AbstractJavaVirtualMachine.java
@@ -114,11 +114,8 @@ public abstract class AbstractJavaVirtualMachine implements JavaVirtualMachine {
             this.jvmConfigurationFromParser = new UnifiedDiarizer();
 
             gcLogFile.stream()
-                    .filter(Objects::nonNull)
-                    .map(String::trim)
-                    .filter(s -> s.length() > 0)
                     .map(this.jvmConfigurationFromParser::diarize)
-                    .filter(completed -> completed)
+                    .filter(Boolean::booleanValue)
                     .findFirst();
 
             this.jvmConfigurationFromParser.fillInKnowns();
